--- submit-web/src/routes/proponent/_proponentLayout/user-management/_userManagementLayout.tsx
@@ -1,8 +1,0 @@
-import { createFileRoute, Outlet } from "@tanstack/react-router";
-
-export const Route = createFileRoute(
-  "/proponent/_proponentLayout/user-management/_userManagementLayout"
-)({
-  meta: () => [{ title: "User Management" }],
-  component: () => <Outlet />,
-});
