+++ submit-web/src/routes/proponent/_proponentLayout/user-management/new-user.tsx
@@ -4,9 +4,12 @@ import { Grid } from "@mui/material";
 import { createFileRoute } from "@tanstack/react-router";
 
 export const Route = createFileRoute(
-  "/proponent/_proponentLayout/user-management/_userManagementLayout/new-user"
+  "/proponent/_proponentLayout/user-management/new-user"
 )({
-  meta: () => [{ title: "Add New User" }],
+  meta: () => [
+    { title: "User Management", path: "/proponent/user-management" },
+    { title: "Add New User", path: "/proponent/user-management/new-user" },
+  ],
   component: () => (
     <PageGrid>
       <Grid item xs={12}>
