+++ submit-web/src/routes/proponent/_proponentLayout/user-management/profile.tsx
@@ -10,7 +10,7 @@ import { createFileRoute } from "@tanstack/react-router";
 import { useAuth } from "react-oidc-context";
 
 export const Route = createFileRoute(
-  "/proponent/_proponentLayout/user-management/_userManagementLayout/profile"
+  "/proponent/_proponentLayout/user-management/profile"
 )({
   component: Profile,
   meta: () => [{ title: "Admin" }],
